+++ app/pages/mongo.vue
@@ -1,7 +1,0 @@
-<script setup lang="ts">
-await navigateTo('/mongo-mgmt', { redirectCode: 301, replace: true })
-</script>
-
-<template>
-  <div />
-</template>